Refactor server commands, new temp textures - 0.0.38
Replace the old monolithic Commands.cs with a new commands/ hierarchy and metadata, adding individual command implementations (Clear, Fill, Help, Info, Kick, Kill, Playerlist, Replace, Spawn, Stop, Tp, etc.).

Other related API and tooling changes:
- Introduce ICmdExecutor.CmdResult.Info and add Console.LogInfo for informational logging.
- Switch Program to use ServerRunner.Instance.
- Improve BlockData1: add MAX_VARIANT constant, use it for masking, and implement ToString(separator).
- Add Vec2Int.MinCorner/MaxCorner helpers.
- Move and enhance world-name validation and password checks into Validation (add WrongWorldNameInfo, IsValidWorldName, disallow whitespace in passwords); remove duplicate IsValidWorldName from Common.
- Bump Core.Version Current to 0.0.38.
- Rename/move some socket/security helpers (concurrent limiter, IEncryptionKey) and other minor server/socket tweaks.
- Update several block texture assets.

Overall this commit modularizes command handling, tightens validation and logging, and includes related refactors and resource updates.

+++ Assets/Assemblies/Socket/Helpers/TrafficLimiter.cs
@@ -5,13 +5,13 @@ namespace Larnix.Socket
 {
     public class TrafficLimiter<T>
     {
-        public readonly uint MAX_TRAFFIC_LOCAL;
-        public readonly uint MAX_TRAFFIC_GLOBAL;
+        public readonly ulong MAX_TRAFFIC_LOCAL;
+        public readonly ulong MAX_TRAFFIC_GLOBAL;
 
-        private readonly Dictionary<T, uint> _localTraffic = new();
-        private uint _globalTraffic = 0;
+        private readonly Dictionary<T, ulong> _localTraffic = new();
+        private ulong _globalTraffic = 0;
 
-        public TrafficLimiter(uint maxTrafficLocal, uint maxTrafficGlobal)
+        public TrafficLimiter(ulong maxTrafficLocal, ulong maxTrafficGlobal)
         {
             MAX_TRAFFIC_LOCAL = maxTrafficLocal;
             MAX_TRAFFIC_GLOBAL = maxTrafficGlobal;
@@ -22,7 +22,7 @@ namespace Larnix.Socket
             if (_globalTraffic >= MAX_TRAFFIC_GLOBAL)
                 return false;
 
-            _localTraffic.TryGetValue(key, out uint localTraffic);
+            _localTraffic.TryGetValue(key, out ulong localTraffic);
             if (localTraffic >= MAX_TRAFFIC_LOCAL)
                 return false;
 
@@ -34,7 +34,7 @@ namespace Larnix.Socket
 
         public void Decrease(T key)
         {
-            if (!_localTraffic.TryGetValue(key, out uint localTraffic))
+            if (!_localTraffic.TryGetValue(key, out ulong localTraffic))
                 throw new InvalidOperationException("Local traffic for this key is already zero.");
 
             localTraffic--;
